pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=5.0.0",
    "flask>=3.1.2",
    "flask-limiter>=4.1.1",
    "pillow>=12.1.0",
    "pyjwt>=2.10.1",
    "pyotp>=2.9.0",
    "pytest>=9.0.2",
    "qrcode>=8.2",
    "qrcode-terminal>=0.8",
]
